# Epico 2/password-api/src/swagger.yaml
openapi: 3.0.3
info:
  title: API de Gestão de Utilizador e Aplicações
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor local
paths:
  /api/register:
    post:
      summary: Registra um novo utilizador
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  minLength: 5
                  maxLength: 20
                  description: Nome de utilizador (5 a 20 caracteres)
                password:
                  type: string
                  minLength: 8
                  maxLength: 32
                  description: Senha do utilizador (8 a 32 caracteres)
              required:
                - username
                - password
      responses:
        '201':
          description: Utilizador criado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                  username:
                    type: string
              example:
                _id: "12345"
                username: "JoaquimZe"
        '400':
          description: Erro na requisição
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - Username e password são obrigatórios
                      - Username precisa de ter entre 5 e 20 caracteres
                      - Password precisa de ter entre 8 e 32 caracteres
                      - Utilizador já existe
              examples:
                missingFields:
                  summary: Campos obrigatórios ausentes
                  value:
                    error: Username e password são obrigatórios
                usernameLength:
                  summary: Username fora dos limites
                  value:
                    error: Username precisa de ter entre 5 e 20 caracteres
                passwordLength:
                  summary: Password fora dos limites
                  value:
                    error: Password precisa de ter entre 8 e 32 caracteres
                userExists:
                  summary: Utilizador já existe
                  value:
                    error: Utilizador já existe
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: Erro interno do servidor
  /api/login:
    post:
      summary: Autentica um utilizador
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: Nome de utilizador
                password:
                  type: string
                  description: Senha do utilizador
              required:
                - username
                - password
      responses:
        '200':
          description: Autenticação bem-sucedida
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    type: object
                    properties:
                      _id:
                        type: string
                      username:
                        type: string
              example:
                token: "jwt.token.aqui"
                user:
                  _id: "12345"
                  username: "JoaquimZe"
        '400':
          description: Erro na requisição
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - Username e password são obrigatórios
                      - Credenciais inválidas
              examples:
                missingFields:
                  summary: Campos obrigatórios ausentes
                  value:
                    error: Username e password são obrigatórios
                invalidCredentials:
                  summary: Credenciais inválidas
                  value:
                    error: Credenciais inválidas
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: Erro interno do servidor
  /api/app:
    post:
      summary: Cria um novo aplicativo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 10
                  maxLength: 40
                  description: Nome do aplicativo (10 a 40 caracteres)
              required:
                - name
      responses:
        '201':
          description: Aplicativo criado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  appid:
                    type: string
                  name:
                    type: string
                  owner:
                    type: string
              example:
                appid: "uuid-aqui"
                name: "MeuApp"
                owner: "12345"
        '400':
          description: Erro na requisição
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - Nome do aplicativo é obrigatório
                      - Nome do aplicativo precisa de ter entre 3 e 50 caracteres
              examples:
                missingName:
                  summary: Nome do aplicativo ausente
                  value:
                    error: Nome do aplicativo é obrigatório
                nameLength:
                  summary: Nome do aplicativo fora dos limites
                  value:
                    error: Nome do aplicativo precisa de ter entre 3 e 50 caracteres
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - Autenticação necessária
              example:
                error: Autenticação necessária
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: Erro interno do servidor
  /api/apps:
    get:
      summary: Lista os aplicativos do utilizador
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de aplicativos retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    appid:
                      type: string
                    name:
                      type: string
                    owner:
                      type: string
                    editors:
                      type: array
                      items:
                        type: string
              example:
                - appid: "uuid-aqui"
                  name: "MeuApp"
                  owner: "12345"
                  editors: []
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - Autenticação necessária
              example:
                error: Autenticação necessária
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: Erro interno do servidor
  /api/password/{appid}:
    post:
      summary: Cria uma nova password para um aplicativo
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: appid
          schema:
            type: string
          required: true
          description: ID único do aplicativo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  minLength: 8
                  maxLength: 32
                  description: Senha do aplicativo (8 a 32 caracteres)
              required:
                - password
      responses:
        '201':
          description: Password criada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  appid:
                    type: string
                  password:
                    type: string
                  createdBy:
                    type: string
              example:
                appid: "uuid-aqui"
                password: "senha12345"
                createdBy: "12345"
        '400':
          description: Erro na requisição
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - Password é obrigatória
                      - Password precisa de ter entre 8 e 32 caracteres
                      - Aplicação não encontrada
              examples:
                missingPassword:
                  summary: Password ausente
                  value:
                    error: Password é obrigatória
                passwordLength:
                  summary: Password fora dos limites
                  value:
                    error: Password precisa de ter entre 8 e 32 caracteres
                appNotFound:
                  summary: Aplicação não encontrada
                  value:
                    error: Aplicação não encontrada
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - Autenticação necessária
              example:
                error: Autenticação necessária
        '403':
          description: Acesso negado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - Acesso negado
              example:
                error: Acesso negado
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: Erro interno do servidor
    put:
      summary: Atualiza a password de um aplicativo
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: appid
          schema:
            type: string
          required: true
          description: ID único do aplicativo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  minLength: 8
                  maxLength: 32
                  description: Senha do aplicativo (8 a 32 caracteres)
              required:
                - password
      responses:
        '200':
          description: Password atualizada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  appid:
                    type: string
                  password:
                    type: string
                  createdBy:
                    type: string
              example:
                appid: "uuid-aqui"
                password: "novasenha12345"
                createdBy: "12345"
        '400':
          description: Erro na requisição
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - Password é obrigatória
                      - Password precisa de ter entre 8 e 32 caracteres
                      - Aplicação não encontrada
                      - Aplicação sem password definida
              examples:
                missingPassword:
                  summary: Password ausente
                  value:
                    error: Password é obrigatória
                passwordLength:
                  summary: Password fora dos limites
                  value:
                    error: Password precisa de ter entre 8 e 32 caracteres
                appNotFound:
                  summary: Aplicação não encontrada
                  value:
                    error: Aplicação não encontrada
                noPassword:
                  summary: Password não definida
                  value:
                    error: Aplicação sem password definida
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - Autenticação necessária
              example:
                error: Autenticação necessária
        '403':
          description: Acesso negado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - Acesso negado
              example:
                error: Acesso negado
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: Erro interno do servidor
    delete:
      summary: Remove a password de um aplicativo
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: appid
          schema:
            type: string
          required: true
          description: ID único do aplicativo
      responses:
        '200':
          description: Password removida com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: Password removida com sucesso
        '400':
          description: Erro na requisição
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - Aplicação não encontrada
                      - Aplicação sem password definida
              examples:
                appNotFound:
                  summary: Aplicação não encontrada
                  value:
                    error: Aplicação não encontrada
                noPassword:
                  summary: Password não definida
                  value:
                    error: Aplicação sem password definida
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - Autenticação necessária
              example:
                error: Autenticação necessária
        '403':
          description: Acesso negado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - Acesso negado
              example:
                error: Acesso negado
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              example:
                error: Erro interno do servidor
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT